requirement.txt
absl-py==2.1.0
addict==2.4.0
asttokens==2.4.1
backcall==0.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
descartes==1.1.0
einops==0.8.0
executing==2.1.0
fire==0.7.0
flash_attn==0.2.0
fonttools==4.54.1
google-auth==2.35.0
google-auth-oauthlib==1.0.0
grpcio==1.67.1
h5py==3.11.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipdb==0.13.13
ipython==8.12.3
jedi==0.19.1
joblib==1.4.2
kiwisolver==1.4.7
llvmlite==0.38.1
loguru==0.7.2
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.7
mmcv-full==1.4.0
mmdet==2.20.0
mpi4py==3.0.3
msgpack==1.1.0
msgpack-numpy==0.4.8
multimethod==1.10
numba==0.55.2
numpy==1.22.4
nuscenes-devkit==1.1.11
oauthlib==3.2.2
opencv-python==4.10.0.84
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==8.4.0
pip==24.2
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.7
Pygments==2.18.0
pyparsing==3.1.4
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
setuptools==59.5.0
Shapely==1.8.5.post1
six==1.16.0
stack-data==0.6.3
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorpack==0.11
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.5.0
toml==0.10.2
tomli==2.0.2
torch==1.10.1+cu113
torchaudio==0.10.1+cu113
torchex==0.1.0
torchpack==0.3.1
torchvision==0.11.2+cu113
tqdm==4.66.6
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.6
wheel==0.44.0
yapf==0.40.1
zipp==3.20.2
